Function is working!
I still have a space at the end of the string, but for now I think this shows the solution well. Initially I was using the if/else statement, but that was because I was having issues passing my parameter to 'var i = x'. I don't know what my problem was initially, but now I see that it should have been working.

If I wanted to take the trailing space off of my solution, I would need to create an if statement within my for loop.